Implemented gears with custom twist profiles
A new `CustomTwistGear` class behaves similarly to `HerringboneGear`, but with twist angle specified by a lambda function in terms of z-position.

The class takes a `twist_function` parameter, which will receive a float between 0 and 1 (representing relative position along the z-axis) and should return a float between 0 and 1 (representing the relative amount of twist).

The resulting gear will have a varying twist between 0 and `twist_angle` along its axis of extrusion.

This allows fancy gear profiles, such as sinusoidal gears.